+++ src/QuickInfo/ImageHtml/ImageQuickInfoSourceProvider.cs
@@ -1,0 +1,19 @@
+﻿using Microsoft.VisualStudio.Language.Intellisense;
+using Microsoft.VisualStudio.Text;
+using Microsoft.VisualStudio.Utilities;
+using System.ComponentModel.Composition;
+
+namespace HtmlTools
+{
+    [Export(typeof(IQuickInfoSourceProvider))]
+    [Name("Image HTML QuickInfo Source")]
+    [Order(Before = "Default Quick Info Presenter")]
+    [ContentType("htmlx")]
+    internal class ImageHtmlQuickInfoSourceProvider : IQuickInfoSourceProvider
+    {
+        public IQuickInfoSource TryCreateQuickInfoSource(ITextBuffer textBuffer)
+        {
+            return new ImageHtmlQuickInfo();
+        }
+    }
+}